Rename back to openstax_rescue_from
