+++ lib/openstax_rescue_from.rb
